DownloadGPXButton component.
